[Enhancement] Allow Custom Size for Textbook Thumbnail
This commit introduces an enhancement to the 'generate_textbook_thumbnail' function in Code 2. Previously, the function generated thumbnails with a fixed size of 220 pixels in width and 300 pixels in height. However, it has now been improved to provide greater flexibility and customization.

The 'generate_textbook_thumbnail' function now accepts an optional 'size' parameter, allowing users to specify the desired dimensions for the generated thumbnail. Users can pass a tuple containing the desired width and height values. For example, to generate a thumbnail with a width of 180 pixels and a height of 250 pixels, the function can be called as follows:

```python
generate_textbook_thumbnail(file, destination, size=(180, 250))